Rename plugin folder
Update plugin name and update folder to opensearch.
