+++ lib/logstash/outputs/opensearch/http_client/manticore_adapter.rb
