+++ lib/logstash/outputs/opensearch/http_client/pool.rb
