+++ lib/logstash/outputs/opensearch/http_client_builder.rb
