+++ lib/logstash/outputs/opensearch/template_manager.rb
