+++ lib/logstash/outputs/opensearch/templates/ecs-disabled/elasticsearch-7x.json
