+++ lib/logstash/plugin_mixins/opensearch/api_configs.rb
