+++ spec/fixtures/_nodes/7x.json
@@ -4,7 +4,7 @@
     "successful" : 3,
     "failed" : 0
   },
-  "cluster_name" : "elasticsearch",
+  "cluster_name" : "opensearch",
   "nodes" : {
     "kVPTh7ZvSgWmTRMy-4YExQ" : {
       "name" : "kVPTh7Z",
