+++ spec/integration/outputs/index_version_spec.rb
@@ -1,8 +1,8 @@
 require_relative "../../../spec/opensearch_spec_helper"
-require "logstash/outputs/elasticsearch"
+require "logstash/outputs/opensearch"
 
 describe "Versioned indexing", :integration => true do
-  require "logstash/outputs/elasticsearch"
+  require "logstash/outputs/opensearch"
 
   let(:es) { get_client }
