+++ spec/integration/outputs/templates_spec.rb
@@ -2,7 +2,7 @@ require_relative "../../../spec/opensearch_spec_helper"
 
 describe "index template expected behavior", :integration => true do
   subject! do
-    require "logstash/outputs/elasticsearch"
+    require "logstash/outputs/opensearch"
     settings = {
       "manage_template" => true,
       "template_overwrite" => true,
