+++ spec/integration/outputs/update_spec.rb
@@ -1,7 +1,7 @@
 require_relative "../../../spec/opensearch_spec_helper"
 
 describe "Update actions without scripts", :integration => true do
-  require "logstash/outputs/elasticsearch"
+  require "logstash/outputs/opensearch"
 
   def get_es_output( options={} )
     settings = {
