+++ spec/unit/outputs/opensearch/http_client/pool_spec.rb
@@ -1,5 +1,5 @@
 require "logstash/devutils/rspec/spec_helper"
-require "logstash/outputs/elasticsearch/http_client"
+require "logstash/outputs/opensearch/http_client"
 require 'cabin'
 
 describe LogStash::Outputs::ElasticSearch::HttpClient::Pool do
@@ -80,27 +80,6 @@ describe LogStash::Outputs::ElasticSearch::HttpClient::Pool do
     let(:ip_address) { "192.168.1.0"}
     let(:port) { 9200 }
 
-    context 'in Elasticsearch 1.x format' do
-      context 'with host and ip address' do
-        let(:publish_address) { "inet[#{host}/#{ip_address}:#{port}]"}
-        it 'should correctly extract the host' do
-          expect(subject.address_str_to_uri(publish_address)).to eq (LogStash::Util::SafeURI.new("#{host}:#{port}"))
-        end
-      end
-      context 'with ip address' do
-        let(:publish_address) { "inet[/#{ip_address}:#{port}]"}
-        it 'should correctly extract the ip address' do
-          expect(subject.address_str_to_uri(publish_address)).to eq (LogStash::Util::SafeURI.new("#{ip_address}:#{port}"))
-        end
-      end
-    end
-
-    context 'in Elasticsearch 2.x-6.x format' do
-      let(:publish_address) { "#{ip_address}:#{port}"}
-      it 'should correctly extract the ip address' do
-        expect(subject.address_str_to_uri(publish_address)).to eq (LogStash::Util::SafeURI.new("//#{ip_address}:#{port}"))
-      end
-    end
 
     context 'in Elasticsearch 7.x'
     context 'with host and ip address' do
